+++ src/util/unicode.cpp
@@ -32,7 +32,7 @@ std::string narrow(const wchar_t *s)
 {
   #ifdef _WIN32
 
-  int slength=wcslen(s);
+  size_t slength=wcslen(s);
   int rlength=WideCharToMultiByte(CP_UTF8, 0, s, slength, NULL, 0, NULL, NULL);
   std::string r(rlength, 0);
   WideCharToMultiByte(CP_UTF8, 0, s, slength, &r[0], rlength, NULL, NULL);
@@ -68,7 +68,7 @@ std::wstring widen(const char *s)
 {
   #ifdef _WIN32
 
-  int slength=strlen(s);
+  size_t slength=strlen(s);
   int rlength=MultiByteToWideChar(CP_UTF8, 0, s, slength, NULL, 0);
   std::wstring r(rlength, 0);
   MultiByteToWideChar(CP_UTF8, 0, s, slength, &r[0], rlength);
@@ -104,7 +104,7 @@ std::string narrow(const std::wstring &s)
 {
   #ifdef _WIN32
 
-  int slength=s.size();
+  size_t slength=s.size();
   int rlength=WideCharToMultiByte(CP_UTF8, 0, &s[0], slength, NULL, 0, NULL, NULL);
   std::string r(rlength, 0);
   WideCharToMultiByte(CP_UTF8, 0, &s[0], slength, &r[0], rlength, NULL, NULL);
@@ -132,7 +132,7 @@ std::wstring widen(const std::string &s)
 {
   #ifdef _WIN32
 
-  int slength=s.size();
+  size_t slength=s.size();
   int rlength=MultiByteToWideChar(CP_UTF8, 0, &s[0], slength, NULL, 0);
   std::wstring r(rlength, 0);
   MultiByteToWideChar(CP_UTF8, 0, &s[0], slength, &r[0], rlength);
